+++ app/jobs/warm_entry_cache_job.rb
@@ -6,12 +6,14 @@ class WarmEntryCacheJob < ApplicationJob
 
   def perform
     category = GetCategory.new(category_entry_id: ENV["CONTENTFUL_DEFAULT_CATEGORY_ENTRY_ID"]).call
-    entries = GetEntriesInCategory.new(category: category).call
+    sections = GetSectionsFromCategory.new(category: category).call
+    steps = sections.map { |section| GetStepsFromSection.new(section: section).call }
 
-    entries.each do |entry|
+    # TODO: Cache category and sections too
+    [steps].flatten.each do |entry|
       store_in_cache(cache: cache, key: "contentful:entry:#{entry.id}", entry: entry)
     end
-  rescue GetEntriesInCategory::RepeatEntryDetected
+  rescue GetStepsFromSection::RepeatEntryDetected
     cache.extend_ttl_on_all_entries
   end