+++ app/services/get_sections_from_category.rb
@@ -1,0 +1,10 @@
+class GetSectionsFromCategory
+  attr_accessor :category
+  def initialize(category:)
+    self.category = category
+  end
+
+  def call
+    category.sections.map { |section| GetEntry.new(entry_id: section.id).call }
+  end
+end