+++ spec/services/create_journey_step_spec.rb
@@ -5,7 +5,7 @@ RSpec.describe CreateJourneyStep do
     context "when the new step is of type step" do
       it "creates a local copy of the new step" do
         journey = create(:journey, :catering)
-        fake_entry = fake_contentful_entry(
+        fake_entry = fake_contentful_step(
           contentful_fixture_filename: "steps/radio-question.json"
         )
 
@@ -61,7 +61,7 @@ RSpec.describe CreateJourneyStep do
     context "when the question is of type 'short_text'" do
       it "sets help_text and options to nil" do
         journey = create(:journey, :catering)
-        fake_entry = fake_contentful_entry(
+        fake_entry = fake_contentful_step(
           contentful_fixture_filename: "steps/short-text-question.json"
         )
 
@@ -72,7 +72,7 @@ RSpec.describe CreateJourneyStep do
 
       it "replaces spaces with underscores" do
         journey = create(:journey, :catering)
-        fake_entry = fake_contentful_entry(
+        fake_entry = fake_contentful_step(
           contentful_fixture_filename: "steps/short-text-question.json"
         )
 
@@ -85,7 +85,7 @@ RSpec.describe CreateJourneyStep do
     context "when the new entry has a body field" do
       it "updates the step with the body" do
         journey = create(:journey, :catering)
-        fake_entry = fake_contentful_entry(
+        fake_entry = fake_contentful_step(
           contentful_fixture_filename: "steps/static-content.json"
         )
 
@@ -103,7 +103,7 @@ process around March.")
     context "when the new entry has a 'primaryCallToAction' field" do
       it "updates the step with the body" do
         journey = create(:journey, :catering)
-        fake_entry = fake_contentful_entry(
+        fake_entry = fake_contentful_step(
           contentful_fixture_filename: "steps/primary-button.json"
         )
 
@@ -118,7 +118,7 @@ process around March.")
     context "when no 'primaryCallToAction' is provided" do
       it "default copy is used for the button" do
         journey = create(:journey, :catering)
-        fake_entry = fake_contentful_entry(
+        fake_entry = fake_contentful_step(
           contentful_fixture_filename: "steps/no-primary-button.json"
         )
 
@@ -133,7 +133,7 @@ process around March.")
     context "when the new entry has an unexpected content model" do
       it "raises an error" do
         journey = create(:journey, :catering)
-        fake_entry = fake_contentful_entry(
+        fake_entry = fake_contentful_step(
           contentful_fixture_filename: "steps/unexpected-contentful-type.json"
         )
 
@@ -144,7 +144,7 @@ process around March.")
       it "raises a rollbar event" do
         journey = create(:journey, :catering)
 
-        fake_entry = fake_contentful_entry(
+        fake_entry = fake_contentful_step(
           contentful_fixture_filename: "steps/unexpected-contentful-type.json"
         )
 
@@ -167,7 +167,7 @@ process around March.")
     context "when the new step has an unexpected step type" do
       it "raises an error" do
         journey = create(:journey, :catering)
-        fake_entry = fake_contentful_entry(
+        fake_entry = fake_contentful_step(
           contentful_fixture_filename: "steps/unexpected-contentful-question-type.json"
         )
 
@@ -178,7 +178,7 @@ process around March.")
       it "raises a rollbar event" do
         journey = create(:journey, :catering)
 
-        fake_entry = fake_contentful_entry(
+        fake_entry = fake_contentful_step(
           contentful_fixture_filename: "steps/unexpected-contentful-question-type.json"
         )