+++ spec/services/get_steps_from_section_spec.rb
@@ -1,14 +1,14 @@
 require "rails_helper"
 
-RSpec.describe GetEntriesInCategory do
+RSpec.describe GetStepsFromSection do
   describe "#call" do
-    it "returns the list of entry objects referenced by the category list" do
-      category = fake_contentful_category(
-        contentful_fixture_filename: "journey-with-multiple-entries.json"
+    it "returns the list of entry objects referenced by the section list" do
+      section = fake_contentful_section(
+        contentful_fixture_filename: "sections/journey-with-multiple-entries-section.json"
       )
-      stub_contentful_category_steps(category: category)
+      stub_contentful_section_steps(sections: [section])
 
-      result = described_class.new(category: category).call
+      result = described_class.new(section: section).call
 
       expect(result).to be_kind_of(Array)
       # INFO: We should test this is a Contentful::Entry however it wasn't
@@ -19,12 +19,12 @@ RSpec.describe GetEntriesInCategory do
 
     context "when the same entry is found twice" do
       it "returns an error message" do
-        category = fake_contentful_category(
-          contentful_fixture_filename: "journey-with-repeat-entries.json"
+        section = fake_contentful_section(
+          contentful_fixture_filename: "sections/journey-with-repeat-entries-section.json"
         )
 
         expect(Rollbar).to receive(:error)
-          .with("A repeated Contentful entry was found in the same journey",
+          .with("A repeated Contentful entry was found in the same section",
             contentful_url: ENV["CONTENTFUL_URL"],
             contentful_space_id: ENV["CONTENTFUL_SPACE"],
             contentful_environment: ENV["CONTENTFUL_ENVIRONMENT"],
@@ -32,8 +32,8 @@ RSpec.describe GetEntriesInCategory do
           .and_call_original
 
         expect {
-          described_class.new(category: category).call
-        }.to raise_error(GetEntriesInCategory::RepeatEntryDetected)
+          described_class.new(section: section).call
+        }.to raise_error(GetStepsFromSection::RepeatEntryDetected)
       end
     end
   end